Load .powconfig as well
This is slightly different than how pow uses .powconfig. It uses it to
configure itself via env variables but puma-dev has a bunch of command
line options to do that instead.

So we still load it because folks use .powconfig as a way to set
variables that are used in all apps, but the POW_* variables won't have
any effect at present.