Replace potentially existing resolver files during install
This is in order to overcome an issue where pow might have files files installed, such as /etc/resolvers/dev.

When using ioutil.WriteFile to a file that is owned by root and made with pow, it won’t replace the contents.

This patch tries to remove the file prior to writing to it, somewhat overcome the shortcomings of ioutil.WriteFile.