Don't return an error code when checking the version
When checking if a program is installed and working correctly in scripts
I usually run the command (i.e. `puma-dev -V`) inside a conditional to
check if we can use it. Since `puma-dev -V` returns a non-zero (1) exit
status for the program, it can't be used like this.

This changes the version command to return a successful exit code so
that it won't block any scripts depending on the exit code of the
command.