`puma-dev -V` exits with status 0 (not 1) (#230)
* Don't return an error code when checking the version

When checking if a program is installed and working correctly in scripts
I usually run the command (i.e. `puma-dev -V`) inside a conditional to
check if we can use it. Since `puma-dev -V` returns a non-zero (1) exit
status for the program, it can't be used like this.

This changes the version command to return a successful exit code so
that it won't block any scripts depending on the exit code of the
command.

* write tests that will fail with a os.Exit(1) on version string

* better error messages

* refactor os.Exit() out into wrapper method

* ok, so flag doesn't like os.Args being modified...

now that we're messing with flag, need to rething #stubOsArgs

* explicitly set flag values that we care about, s/StubFlagArgs/StubCommandLineArgs/g

* ensure that all flags are reset to their default values before stubbing

* clarify StubCommandLineArgs usage, loosen -V regex for HEAD

travis: "Version: devel (devel +363bcd0068 Wed Feb 12 18:22:50 2020 +0000)\n"

* wrap comments at around 80 chars

* had a false positive due to type checking ok

* assertions are more revealing

* trying to look at coverage

* reset flags to default

* ok, proceed with the reset-to-default-if-stubbed strategy.

because puma-dev mixes flags and commandline arguments, we're abusing os.Args in our tests. we can't just nuke all the flags, because then the flags that are passed in for testing are broken (e.g. we get no coverage). so, we need to record which commandline args have been provide that actually override the cached flag values, and surgically reset them.

* refactor StubCommandLineArgs into a variadic function

* test early exit for command

* alphabetize

* who tests the testers?

* support -flag=val as well as -flag

* refactor tuple return into a struct for better readability

* don't need to check err type, we know it's nil for a clean exit

Co-authored-by: André Ligné <hello@bzf.se>