Fix: sharing app instances with different hostnames not working (#281)
This fixes a problem where it fails to reuse the same instance of the app
when using multiple hostnames with different names (multiple symlinks pointing
to the same folder).

The bug was introduced in #270. The patch is based on adding the checksum in
all cases when a symlink is involved, not only when the names don't match. If not,
it can happen that it's not added if the name matches for the first execution, which
results in not finding the app instance later, resulting in an error "There is already
a server bound to..."

Fixes #280